requirements.txt
Django==5.1.1
python-telegram-bot==22.0
python-dotenv==1.1.0
asgiref==3.8.1